Module google_serviceregistryalpha::api

Structs

Endpoint
There is no detailed description.

EndpointDeleteCall
Deletes an endpoint.

EndpointEndpointVisibility
There is no detailed description.

EndpointGetCall
Gets an endpoint.

EndpointInsertCall
Creates an endpoint.

EndpointListCall
Lists endpoints for a project.

EndpointMethods
A builder providing access to all methods supported on endpoint resources. It is not used directly, but through the ServiceRegistry hub.

EndpointPatchCall
Updates an endpoint. This method supports patch semantics.

EndpointUpdateCall
Updates an endpoint.

EndpointsListResponse
A response containing a partial list of Endpoints and a page token used to build the next request if the request has been truncated. Next available tag: 6

Operation
An Operation resource, used to manage asynchronous API requests.

OperationError
[Output Only] If errors are generated during processing of the operation, this field will be populated.

OperationErrorErrors
[Output Only] The array of errors encountered while processing this operation.

OperationGetCall
Gets information about a specific operation.

OperationListCall
Lists all operations for a project.

OperationMethods
A builder providing access to all methods supported on operation resources. It is not used directly, but through the ServiceRegistry hub.

OperationWarnings
[Output Only] If warning messages are generated during processing of the operation, this field will be populated.

OperationWarningsData
[Output Only] Metadata about this warning in key: value format. For example: "data": [ { "key": "scope", "value": "zones/us-east1-d" }

OperationsListResponse
A response containing a partial list of operations and a page token used to build the next request if the request has been truncated.

ServiceRegistry
Central instance to access all ServiceRegistry related resource activities

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.