Struct google_serviceregistryalpha::Endpoint

pub struct Endpoint {
    pub description: Option<String>,
    pub address: Option<String>,
    pub visibility: Option<EndpointEndpointVisibility>,
    pub id: Option<String>,
    pub state: Option<String>,
    pub fingerprint: Option<String>,
    pub creation_timestamp: Option<String>,
    pub port: Option<i32>,
    pub self_link: Option<String>,
    pub name: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert endpoints (request)

delete endpoints (none)

update endpoints (request)

get endpoints (response)

list endpoints (none)

patch endpoints (request)

Fields

description

An optional user-provided description of the endpoint.

address

A user-provided address of the service represented by this endpoint. This can be an IPv4 or IPv6 address, or a hostname.

visibility

The DNS Integration configuration for this endpoint. This must be a list of fully-qualified URLs to Compute Engine networks.

id

[Output Only] Unique identifier for the resource; defined by the server.

state

[Output Only] The current state of the endpoint, as determined by the system.

fingerprint

Supply the fingerprint value for update requests. The fingerprint value is generated by the server and ensures optimistic concurrency (so that only one update can be performed at a time). The fingerprint changes after each update.

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

port

An optional user-provided port of the service represented by this endpoint.

self_link

[Output Only] Self link for the endpoint.

name

A user-provided name of the endpoint, which must be unique within the project. The name must comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression a-z? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.