Struct google_serviceregistryalpha::EndpointMethods

pub struct EndpointMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on endpoint resources. It is not used directly, but through the ServiceRegistry hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_serviceregistryalpha as serviceregistryalpha;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use serviceregistryalpha::ServiceRegistry;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = ServiceRegistry::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.endpoints();

Methods

impl<'a, C, A> EndpointMethods<'a, C, A>

fn insert(&self, request: Endpoint, project: &str) -> EndpointInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates an endpoint.

Arguments

request - No description provided.

project - The project ID for this request.

fn delete(&self, project: &str, endpoint: &str) -> EndpointDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an endpoint.

Arguments

project - The project ID for this request.

endpoint - The name of the endpoint for this request.

fn update(&self, request: Endpoint, project: &str, endpoint: &str) -> EndpointUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an endpoint.

Arguments

request - No description provided.

project - The project ID for this request.

endpoint - The name of the endpoint for this request.

fn patch(&self, request: Endpoint, project: &str, endpoint: &str) -> EndpointPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an endpoint. This method supports patch semantics.

Arguments

request - No description provided.

project - The project ID for this request.

endpoint - The name of the endpoint for this request.

fn list(&self, project: &str) -> EndpointListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists endpoints for a project.

Arguments

project - The project ID for this request.

fn get(&self, project: &str, endpoint: &str) -> EndpointGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets an endpoint.

Arguments

project - The project ID for this request.

endpoint - The name of the endpoint for this request.